# gvb_server/docs/swagger.yaml
basePath: /
definitions:
  advert_api.AdvertRequest:
    properties:
      href:
        description: 跳转链接
        type: string
      images:
        description: 图片
        type: string
      is_show:
        description: 是否展示
        type: boolean
      title:
        description: 显示的标题
        type: string
    required:
    - href
    - images
    - title
    type: object
  common.ListResponse-models_AdvertModel:
    properties:
      count:
        type: integer
      list:
        $ref: '#/definitions/models.AdvertModel'
    type: object
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  models.AdvertModel:
    properties:
      created_at:
        description: 创建时间
        type: string
      href:
        description: 跳转链接
        type: string
      id:
        description: 主键ID
        type: integer
      images:
        description: 图片
        type: string
      is_show:
        description: 是否展示
        type: boolean
      title:
        description: 显示的标题
        type: string
    type: object
  models.RemoveRequest:
    properties:
      id_list:
        items:
          type: integer
        type: array
    type: object
host: 127.0.0.1:8080
info:
  contact: {}
  description: gvb_server第一版的API文档
  title: gvb_server API文档
  version: "1.0"
paths:
  /api/adverts:
    delete:
      description: 通过id列表的形式一次性删除多个或者一个广告
      parameters:
      - description: 广告ID列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.RemoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 删除广告
      tags:
      - 广告管理
    get:
      description: 可以分页查询广告列表
      parameters:
      - description: 标签 form:"key" 表示该字段可以从表单数据中获取名为 "key" 的值。
        in: query
        name: key
        type: string
      - description: 标签 form:"limit" 表示该字段可以从表单数据中获取名为 "limit" 的值。
        in: query
        name: limit
        type: integer
      - description: 标签 form:"page" 表示该字段可以从表单数据中获取名为 "page" 的值。
        in: query
        name: page
        type: integer
      - description: 标签 form:"sort" 表示该字段可以从表单数据中获取名为 "sort" 的值。
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/common.ListResponse-models_AdvertModel'
              type: object
      summary: 查询广告
      tags:
      - 广告管理
    post:
      description: 关于广告的创建的API，拥有判重逻辑
      parameters:
      - description: 表示多个参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/advert_api.AdvertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 创建广告
      tags:
      - 广告管理
  /api/adverts/{id}:
    put:
      description: 查询ID，来修改该广告的内容
      parameters:
      - description: 广告的一些参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/advert_api.AdvertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 修改广告
      tags:
      - 广告管理
swagger: "2.0"